+++ sharding-core/src/main/java/io/shardingsphere/core/parsing/antlr/sql/ddl/ColumnDefinition.java
@@ -19,7 +19,6 @@ package io.shardingsphere.core.parsing.antlr.sql.ddl;
 
 import lombok.Getter;
 import lombok.Setter;
-import lombok.ToString;
 
 /**
  * Column definition.
@@ -28,7 +27,6 @@ import lombok.ToString;
  */
 @Getter
 @Setter
-@ToString
 public final class ColumnDefinition {
     
     private String name;
@@ -41,19 +39,10 @@ public final class ColumnDefinition {
     
     private ColumnPosition position;
     
-    private String oldName;
-    
-    private boolean isAdd;
-    
     public ColumnDefinition(final String name, final String type, final Integer length, final boolean primaryKey) {
         this.name = name;
         this.type = type;
         this.length = length;
         this.primaryKey = primaryKey;
-    }
-    
-    public ColumnDefinition(final String name, final String oldName) {
-        this.name = name;
-        this.oldName = oldName;
     }
 }