+++ sharding-core/src/main/java/io/shardingsphere/core/parsing/parser/sql/ddl/DDLStatement.java
@@ -22,7 +22,6 @@ import io.shardingsphere.core.parsing.lexer.token.DefaultKeyword;
 import io.shardingsphere.core.parsing.lexer.token.Keyword;
 import io.shardingsphere.core.parsing.lexer.token.TokenType;
 import io.shardingsphere.core.parsing.parser.sql.AbstractSQLStatement;
-import lombok.ToString;
 
 import java.util.Arrays;
 import java.util.Collection;
@@ -32,7 +31,6 @@ import java.util.Collection;
  *
  * @author zhangliang
  */
-@ToString(callSuper = true)
 public class DDLStatement extends AbstractSQLStatement {
     
     private static final Collection<Keyword> PRIMARY_STATEMENT_PREFIX = Arrays.<Keyword>asList(DefaultKeyword.CREATE, DefaultKeyword.ALTER, DefaultKeyword.DROP, DefaultKeyword.TRUNCATE);