+++ sharding-core/src/test/java/io/shardingsphere/core/routing/type/hint/DatabaseHintRoutingEngineTest.java
@@ -50,7 +50,8 @@ public final class DatabaseHintRoutingEngineTest {
         shardingRuleConfig.getTableRuleConfigs().add(tableRuleConfig);
         shardingRuleConfig.setDefaultDatabaseShardingStrategyConfig(new HintShardingStrategyConfiguration(new OrderDatabaseHintShardingAlgorithm()));
         ShardingRule shardingRule = new ShardingRule(shardingRuleConfig, Arrays.asList("ds_0", "ds_1"));
-        databaseHintRoutingEngine = new DatabaseHintRoutingEngine(shardingRule.getShardingDataSourceNames().getDataSourceNames(), (HintShardingStrategy) shardingRule.getDefaultDatabaseShardingStrategy());
+        databaseHintRoutingEngine = new DatabaseHintRoutingEngine(
+                shardingRule.getShardingDataSourceNames().getDataSourceNames(), (HintShardingStrategy) shardingRule.getDefaultDatabaseShardingStrategy());
     }
     
     @After